Move StandalonePersistRepositoryConfiguration
